+++ src/render/utils/index.js
@@ -1,10 +1,15 @@
 const { classNameCase, formatDescription, internalCase } = require("../../common");
-const { getInlineParseContent } = require("../../schema");
+const { getComponentByRef } = require("../../components");
+const { getInlineParseContent, getParseContent, parseSchema } = require("../../schema");
 
 module.exports = {
   formatDescription,
   internalCase,
   classNameCase,
+  getInlineParseContent,
+  getParseContent,
+  getComponentByRef,
+  parseSchema,
   fmtToJSDocLine: require("./fmtToJSDocLine"),
   _: require("lodash"),
 };