+++ src/modelTypes.js
@@ -1,27 +1,14 @@
-const _ = require("lodash");
 const { formatters } = require("./typeFormatters");
 const { checkAndRenameModelName } = require("./modelNames");
-const { formatDescription } = require("./common");
 const { config } = require("./config");
 const { getTypeData } = require("./components");
 
-const CONTENT_KEYWORD = "__CONTENT__";
-
-const contentWrapersByTypeIdentifier = {
-  enum: `{\r\n${CONTENT_KEYWORD} \r\n }`,
-  interface: `{\r\n${CONTENT_KEYWORD}}`,
-  type: `= ${CONTENT_KEYWORD}`,
-};
-
-const getModelType = (typeInfo) => {
-  let { typeIdentifier, name: originalName, content, type, description } = getTypeData(typeInfo);
+const prepareModelType = (typeInfo) => {
+  const typeData = getTypeData(typeInfo);
+  let { typeIdentifier, name: originalName, content, type, description } = typeData;
 
   if (config.generateUnionEnums && typeIdentifier === "enum") {
     typeIdentifier = "type";
-  }
-
-  if (!contentWrapersByTypeIdentifier[typeIdentifier]) {
-    throw new Error(`${typeIdentifier} - type identifier is unknown for this utility`);
   }
 
   const resultContent = formatters[type] ? formatters[type](content) : content;
@@ -30,17 +17,13 @@ const getModelType = (typeInfo) => {
   return {
     typeIdentifier,
     name,
-    rawContent: resultContent,
-    description: formatDescription(description),
-    content: _.replace(
-      contentWrapersByTypeIdentifier[typeIdentifier],
-      CONTENT_KEYWORD,
-      resultContent,
-    ),
+    description,
+    rawContent: content,
+    content: resultContent,
+    typeData,
   };
 };
 
 module.exports = {
-  getModelType,
-  checkAndRenameModelName,
+  prepareModelType,
 };