+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # next release  
 
 BREAKING_CHANGES:  
-- Migrate from [mustache](https://mustache.github.io/) template engine to [ETA](https://eta.js.org/) template engine.  
+- Migrate from [mustache](https://mustache.github.io/) template engine to [ETA](https://eta.js.org/) template engine. (Thanks @Fl0pZz)  
 
 Features:  
 - `--modular` option. Allows to generate api class per module name.   
   Example: [here](./tests/spec/modular)   
-- new templates on [ETA](https://eta.js.org/) (enhanced EJS) which can improve your templates!   
+- new templates on [ETA](https://eta.js.org/) (enhanced EJS) which can improve your templates! (Thanks @Fl0pZz)   
   [ETA extension for Visual Studio Code](https://marketplace.visualstudio.com/items?itemName=shadowtime2000.eta-vscode) (thanks @shadowtime2000)  
   Also moved out to templates:
     - `procedure-call.eta` (request api method template)